+++ app/forms/__init__.py
@@ -1,0 +1,1 @@
+from .auth import LoginForm  # noqa
