+++ app/forms/auth.py
@@ -4,6 +4,6 @@ from wtforms.validators import DataRequired
 
 
 class LoginForm(FlaskForm):
-    user_id = StringField("Name", [DataRequired()])
+    user_name = StringField("Name", [DataRequired()])
     password = PasswordField("Password", [DataRequired()])
     submit = SubmitField("Login")
