+++ app/templates/login.html
@@ -9,8 +9,8 @@
     <div class="card-body">
       <form class="form" role="form" action="{{ url_for('auth.login') }}" method="post">
         <div class="form-group">
-          {{form.user_id.label}}
-          {{form.user_id(class='form-control')}}
+          {{form.user_name.label}}
+          {{form.user_name(class='form-control')}}
         </div>
         <div class="form-group">
           {{form.password.label}}
