+++ app/views/__init__.py
@@ -1,0 +1,6 @@
+from .auth import auth_blueprint  # noqa
+from .main import main_blueprint  # noqa
+from .account import account_blueprint  # noqa
+from .product import product_blueprint  # noqa
+from .reseller import reseller_blueprint  # noqa
+from .user import user_blueprint  # noqa
