+++ app/views/auth.py
@@ -2,7 +2,7 @@ from flask import Blueprint, render_template, url_for, redirect, flash, request
 from flask_login import login_user, logout_user, login_required
 
 from ..models import User
-from .forms import LoginForm
+from ..forms import LoginForm
 
 auth_blueprint = Blueprint("auth", __name__)
 
@@ -17,7 +17,7 @@ def login():
             flash("Login successful.", "success")
             return redirect(url_for("main.index"))
         flash("Wrong user ID or password.", "danger")
-    return render_template("auth/login.html", form=form)
+    return render_template("login.html", form=form)
 
 
 @auth_blueprint.route("/logout")
