+++ app/views/main.py
@@ -1,0 +1,9 @@
+from flask import render_template, Blueprint
+
+
+main_blueprint = Blueprint('main', __name__)
+
+
+@main_blueprint.route('/')
+def index():
+    return render_template('base.html')
