+++ app/views/user.py
@@ -1,0 +1,9 @@
+from flask import render_template, Blueprint
+
+
+user_blueprint = Blueprint('user', __name__)
+
+
+@user_blueprint.route("/user_edit")
+def user_edit():
+    return render_template("user_edit.html")
